Move `cuda::latch` to its own file
