+++ libcudacxx/test/libcudacxx/std/thread/thread.latch/max.pass.cpp
@@ -11,6 +11,7 @@
 
 // <cuda/std/latch>
 
+#include <cuda/latch>
 #include <cuda/std/latch>
 
 #include "test_macros.h"
