+++ libcudacxx/test/libcudacxx/std/thread/thread.latch/try_wait.pass.cpp
@@ -11,6 +11,7 @@
 
 // <cuda/std/latch>
 
+#include <cuda/latch>
 #include <cuda/std/cassert>
 #include <cuda/std/latch>
